refactor: move entities to the modules folder
