+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { DataBaseModule } from './database/database.module';
-import { UserModule } from './user/user.module';
-import { ArtistModule } from './artist/artist.module';
-import { TrackModule } from './track/track.module';
-import { AlbumModule } from './album/album.module';
+import { UserModule } from './modules/user/user.module';
+import { ArtistModule } from './modules/artist/artist.module';
+import { TrackModule } from './modules/track/track.module';
+import { AlbumModule } from './modules/album/album.module';
 import { FavoritesModule } from './modules/favorites/favorites.module';
 
 @Module({
