+++ src/modules/artist/dto/create-artist.dto.ts
